requirements.txt
streamlit==1.31.1
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
plotly==5.14.0
google-generativeai==0.3.0
joblib==1.3.1
sqlalchemy==2.0.20
